Move prompt_user_device_timeout to dedicated gui_dialogs module
Cleaner separation of concerns: dialog helpers live in their own
module rather than inside ActionManager which manages action scripts.
